pyproject.toml
[tool.poetry]
name = "mvc-flask"
version = "2.3.1"
description = "turn standard Flask into mvc"
authors = ["Marcus Pereira <marcus@negros.dev>"]

readme = "README.md"
license = "MIT"
keywords = ["flask", "flask extensions", "flask mvc", "flask-mvc", "mvc flask", "flask design pattern"]

repository = "https://github.com/marcuxyz/mvc_flask"
homepage = "https://github.com/marcuxyz/mvc_flask"


[tool.poetry.dependencies]
python = ">=3.7,<4.0.0"
Flask = "^2.0.1"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
Authlib = "^1.0.1"
requests = "^2.28.1"
libsass = "^0.21.0"
python-dotenv = "^0.20.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
ward = "^0.65.0-beta.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
